// src/models/TigResultado.ts
export interface TigResultado {
    diametroEletrodo?: string;
    diametroAdicao?: string;
    velocidadeSoldagem?: string;
    correnteSoldaPasseRaiz?: string;
    correnteSoldaOutrosPasses?: string;
    corrente?: string;
    tensao?: string;
    vazaoArgonio?: string;
    vazaoHelio?: string;
    numPasses?: string;
    tipoJunta?: string;
}